+++ apps/server/src/modules/tool/service/validation/common-tool-validation.service.spec.ts
@@ -144,56 +144,4 @@ describe('CommonToolValidationService', () => {
 			});
 		});
 	});
-
-	describe('isClientIdUnique is called', () => {
-		describe('when config is not of type oauth2', () => {
-			it('should return without error', async () => {
-				const externalToolDO: ExternalToolDO = externalToolDOFactory.build();
-
-				const func = () => service.isClientIdUnique(externalToolDO);
-
-				await expect(func()).resolves.not.toThrow();
-			});
-		});
-
-		describe('when config is of type oauth2', () => {
-			it('should call the service', async () => {
-				const expectedClientId = 'clientId';
-				const externalToolDO: ExternalToolDO = externalToolDOFactory
-					.withOauth2Config({ clientId: expectedClientId })
-					.build();
-
-				await service.isClientIdUnique(externalToolDO);
-
-				expect(externalToolService.findExternalToolByOAuth2ConfigClientId).toHaveBeenCalledWith(expectedClientId);
-			});
-		});
-
-		describe('when tool name does not have changed', () => {
-			it('should return true', async () => {
-				const externalToolDO: ExternalToolDO = externalToolDOFactory.withOauth2Config().buildWithId();
-				externalToolService.findExternalToolByOAuth2ConfigClientId.mockResolvedValue(externalToolDO);
-
-				const result: boolean = await service.isClientIdUnique(externalToolDO);
-
-				expect(result).toBeTruthy();
-			});
-		});
-
-		describe('when another tool with the name already exists', () => {
-			it('should return false', async () => {
-				const externalToolDO: ExternalToolDO = externalToolDOFactory
-					.withOauth2Config({ clientId: 'sameClientId' })
-					.build();
-				const existingExternalToolDO: ExternalToolDO = externalToolDOFactory
-					.withOauth2Config({ clientId: 'sameClientId' })
-					.buildWithId();
-				externalToolService.findExternalToolByOAuth2ConfigClientId.mockResolvedValue(existingExternalToolDO);
-
-				const result: boolean = await service.isClientIdUnique(externalToolDO);
-
-				expect(result).toBeFalsy();
-			});
-		});
-	});
 });
